+++ traffic_portal/app/src/common/modules/table/deliveryServiceRequests/TableDeliveryServiceRequestsController.js
@@ -17,7 +17,70 @@
  * under the License.
  */
 
-var TableDeliveryServicesRequestsController = function (dsRequests, $scope, $state, $uibModal, $anchorScroll, $q, $location, dateUtils, locationUtils, typeService, deliveryServiceService, deliveryServiceRequestService, messageModel, propertiesModel, userModel) {
+var TableDeliveryServicesRequestsController = function (tableName, dsRequests, $scope, $state, $uibModal, $anchorScroll, $q, $location, $document, dateUtils, locationUtils, typeService, deliveryServiceService, deliveryServiceRequestService, messageModel, propertiesModel, userModel) {
+
+	/**
+	 * Gets value to display a default tooltip.
+	 */
+	function defaultTooltip(params) {
+		return params.value;
+	}
+
+	/**
+	 * Formats the contents of a 'createdAt' and 'lastUpdated' column cell as "relative to now".
+	 */
+	function dateCellFormatter(params) {
+		return params.value ? dateUtils.getRelativeTime(params.value) : params.value;
+	}
+
+	columns = [
+		{
+			headerName: "Delivery Service",
+			field: "deliveryService.xmlId",
+			hide: false
+		},
+		{
+			headerName: "Type",
+			field: "changeType",
+			hide: false
+		},
+		{
+			headerName: "Status",
+			field: "status",
+			hide: false
+		},
+		{
+			headerName: "Author",
+			field: "author",
+			hide: false
+		},
+		{
+			headerName: "Assignee",
+			field: "assignee",
+			hide: false
+		},
+		{
+			headerName: "Last Edited By",
+			field: "lastEditedBy",
+			hide: true
+		},
+		{
+			headerName: "Last Updated",
+			field: "lastUpdated",
+			hide: true,
+			filter: "agDateColumnFilter",
+			tooltip: dateCellFormatter,
+			valueFormatter: dateCellFormatter
+		},
+		{
+			headerName: "Created",
+			field: "createdAt",
+			hide: false,
+			filter: "agDateColumnFilter",
+			tooltip: dateCellFormatter,
+			valueFormatter: dateCellFormatter
+		}
+	];
 
 	var createComment = function (request, placeholder) {
 		var params = {
@@ -52,32 +115,166 @@ var TableDeliveryServicesRequestsController = function (dsRequests, $scope, $sta
 	$scope.PENDING = 3;
 	$scope.COMPLETE = 4;
 
-	$scope.dsRequests = dsRequests;
+	/** All of the ds requests - createdAt and lastUpdated fields converted to actual Date */
+	$scope.dsRequests = dsRequests.map(
+		function(x) {
+			x.createdAt = x.createdAt ? new Date(x.createdAt.replace("+00", "Z")) : x.createdAt;
+			x.lastUpdated = x.lastUpdated ? new Date(x.lastUpdated.replace("+00", "Z")) : x.lastUpdated;
+		});
 
-	$scope.getRelativeTime = dateUtils.getRelativeTime;
+	$scope.quickSearch = '';
 
-	$scope.refresh = function () {
-		$state.reload(); // reloads all the resolves for the view
+	$scope.pageSize = 100;
+
+	$scope.mouseDownSelectionText = "";
+
+	/** Options, configuration, data and callbacks for the ag-grid table. */
+	$scope.gridOptions = {
+		onCellMouseDown: function() {
+			$scope.mouseDownSelectionText = window.getSelection().toString();
+		},
+		onRowClicked: function(params) {
+			const selection = window.getSelection().toString();
+			if(selection === "" || selection === $scope.mouseDownSelectionText) {
+				let path = '/delivery-service-requests/' + params.data.id + '?type=';
+				typeService.getType(params.data.deliveryService.typeId)
+					.then(function (result) {
+						path += result.name;
+						locationUtils.navigateToPath(path);
+					});
+				$scope.$apply();
+			}
+			$scope.mouseDownSelectionText = "";
+		},
+		columnDefs: columns,
+		enableCellTextSelection: true,
+		suppressMenuHide: true,
+		multiSortKey: 'ctrl',
+		alwaysShowVerticalScroll: true,
+		defaultColDef: {
+			filter: true,
+			sortable: true,
+			resizable: true,
+			tooltip: defaultTooltip
+		},
+		rowClassRules: {
+			'draft-request': function(params) {
+				return params.data.status === 'draft';
+			},
+			'submitted-request': function(params) {
+				return params.data.status === 'submitted';
+			},
+			'pending-request': function(params) {
+				return params.data.status === 'pending';
+			},
+			'completed-request': function(params) {
+				return params.data.status === 'complete';
+			},
+			'rejected-request': function(params) {
+				return params.data.status === 'rejected';
+			},
+		},
+		rowData: dsRequests,
+		pagination: true,
+		paginationPageSize: $scope.pageSize,
+		rowBuffer: 0,
+		onColumnResized: function(params) {
+			localStorage.setItem(tableName + "_table_columns", JSON.stringify($scope.gridOptions.columnApi.getColumnState()));
+		},
+		tooltipShowDelay: 500,
+		allowContextMenuWithControlKey: true,
+		preventDefaultOnContextMenu: true,
+		onCellContextMenu: function(params) {
+			$scope.showMenu = true;
+			$scope.menuStyle.left = String(params.event.clientX) + "px";
+			$scope.menuStyle.top = String(params.event.clientY) + "px";
+			$scope.menuStyle.bottom = "unset";
+			$scope.menuStyle.right = "unset";
+			$scope.$apply();
+			const boundingRect = document.getElementById("context-menu").getBoundingClientRect();
+
+			if (boundingRect.bottom > window.innerHeight){
+				$scope.menuStyle.bottom = String(window.innerHeight - params.event.clientY) + "px";
+				$scope.menuStyle.top = "unset";
+			}
+			if (boundingRect.right > window.innerWidth) {
+				$scope.menuStyle.right = String(window.innerWidth - params.event.clientX) + "px";
+				$scope.menuStyle.left = "unset";
+			}
+			$scope.request = params.data;
+			$scope.$apply();
+		},
+		onColumnVisible: function(params) {
+			if (params.visible){
+				return;
+			}
+			for (let column of params.columns) {
+				if (column.filterActive) {
+					const filterModel = $scope.gridOptions.api.getFilterModel();
+					if (column.colId in filterModel) {
+						delete filterModel[column.colId];
+						$scope.gridOptions.api.setFilterModel(filterModel);
+					}
+				}
+			}
+		},
+		colResizeDefault: "shift"
+	};
+
+	/** This is used to position the context menu under the cursor. */
+	$scope.menuStyle = {
+		left: 0,
+		top: 0,
+	};
+
+	/** Toggles the visibility of a column that has the ID provided as 'col'. */
+	$scope.toggleVisibility = function(col) {
+		const visible = $scope.gridOptions.columnApi.getColumn(col).isVisible();
+		$scope.gridOptions.columnApi.setColumnVisible(col, !visible);
+	};
+
+	/** Downloads the table as a CSV */
+	$scope.exportCSV = function() {
+		const params = {
+			allColumns: true,
+			fileName: "delivery_service_requests.csv",
+		};
+		$scope.gridOptions.api.exportDataAsCsv(params);
+	}
+
+	$scope.onQuickSearchChanged = function() {
+		$scope.gridOptions.api.setQuickFilter($scope.quickSearch);
+		localStorage.setItem(tableName + "_quick_search", $scope.quickSearch);
+	};
+
+	$scope.onPageSizeChanged = function() {
+		const value = Number($scope.pageSize);
+		$scope.gridOptions.api.paginationSetPageSize(value);
+		localStorage.setItem(tableName + "_page_size", value);
+	};
+
+	$scope.clearTableFilters = function() {
+		// clear the quick search
+		$scope.quickSearch = '';
+		$scope.onQuickSearchChanged();
+		// clear any column filters
+		$scope.gridOptions.api.setFilterModel(null);
 	};
 
 	$scope.fulfillable = function (request) {
-		return request.status == 'submitted';
+		return request && request.status == 'submitted';
 	};
 
 	$scope.rejectable = function (request) {
-		return request.status == 'submitted';
+		return request && request.status == 'submitted';
 	};
 
 	$scope.completeable = function (request) {
-		return request.status == 'pending';
+		return request && request.status == 'pending';
 	};
 
 	$scope.open = function (request) {
-		return (request.status == 'draft' || request.status == 'submitted');
-	};
-
-	$scope.closed = function (request) {
-		return (request.status == 'rejected' || request.status == 'complete');
+		return request && (request.status == 'draft' || request.status == 'submitted');
 	};
 
 	$scope.assignRequest = function (request, assign, $event) {
@@ -247,15 +444,6 @@ var TableDeliveryServicesRequestsController = function (dsRequests, $scope, $sta
 		});
 	};
 
-	$scope.editDeliveryServiceRequest = function (request) {
-		var path = '/delivery-service-requests/' + request.id + '?type=';
-		typeService.getType(request.deliveryService.typeId)
-			.then(function (result) {
-				path += result.name;
-				locationUtils.navigateToPath(path);
-			});
-	};
-
 	$scope.fulfillRequest = function (request, $event) {
 		$event.stopPropagation(); // this kills the click event so it doesn't trigger anything else
 		var path = '/delivery-service-requests/' + request.id + '?type=';
@@ -266,32 +454,88 @@ var TableDeliveryServicesRequestsController = function (dsRequests, $scope, $sta
 			});
 	};
 
+	$scope.refresh = function () {
+		$state.reload(); // reloads all the resolves for the view
+	};
+
+	/**** Initialization code, including loading user columns from localstorage ****/
 	angular.element(document).ready(function () {
-		var dsRequestsTable = $('#dsRequestsTable').dataTable({
-			"paging": false,
-			"dom": '<"filter-checkbox">frtip',
-			"columnDefs": [
-				{'orderable': false, 'targets': 7}
-			],
-			"aaSorting": []
-		});
-		$('div.filter-checkbox').html('<input id="showClosed" type="checkbox"><label for="showClosed">&nbsp;&nbsp;Show Closed</label>');
-
-		// only show "open" ds requests on render
-		dsRequestsTable.fnFilter('draft|submitted|pending', 2, true, false);
-
-		$('#showClosed').click(function () {
-			var checked = $('#showClosed').is(':checked');
-			if (checked) {
-				dsRequestsTable.fnFilter('', 2, true, false);
+		try {
+			// need to create the show/hide column checkboxes and bind to the current visibility
+			const colstates = JSON.parse(localStorage.getItem(tableName + "_table_columns"));
+			if (colstates) {
+				if (!$scope.gridOptions.columnApi.setColumnState(colstates)) {
+					console.error("Failed to load stored column state: one or more columns not found");
+				}
 			} else {
-				dsRequestsTable.fnFilter('draft|submitted|pending', 2, true, false);
+				$scope.gridOptions.api.sizeColumnsToFit();
+			}
+		} catch (e) {
+			console.error("Failure to retrieve required column info from localStorage (key=" + tableName + "_table_columns):", e);
+		}
+
+		try {
+			const filterState = JSON.parse(localStorage.getItem(tableName + "_table_filters")) || {};
+			$scope.gridOptions.api.setFilterModel(filterState);
+		} catch (e) {
+			console.error("Failure to load stored filter state:", e);
+		}
+
+		$scope.gridOptions.api.addEventListener("filterChanged", function() {
+			localStorage.setItem(tableName + "_table_filters", JSON.stringify($scope.gridOptions.api.getFilterModel()));
+		});
+
+		try {
+			const sortState = JSON.parse(localStorage.getItem(tableName + "_table_sort"));
+			$scope.gridOptions.api.setSortModel(sortState);
+		} catch (e) {
+			console.error("Failure to load stored sort state:", e);
+		}
+
+		try {
+			$scope.quickSearch = localStorage.getItem(tableName + "_quick_search");
+			$scope.gridOptions.api.setQuickFilter($scope.quickSearch);
+		} catch (e) {
+			console.error("Failure to load stored quick search:", e);
+		}
+
+		try {
+			const ps = localStorage.getItem(tableName + "_page_size");
+			if (ps && ps > 0) {
+				$scope.pageSize = Number(ps);
+				$scope.gridOptions.api.paginationSetPageSize($scope.pageSize);
+			}
+		} catch (e) {
+			console.error("Failure to load stored page size:", e);
+		}
+
+		$scope.gridOptions.api.addEventListener("sortChanged", function() {
+			localStorage.setItem(tableName + "_table_sort", JSON.stringify($scope.gridOptions.api.getSortModel()));
+		});
+
+		$scope.gridOptions.api.addEventListener("columnMoved", function() {
+			localStorage.setItem(tableName + "_table_columns", JSON.stringify($scope.gridOptions.columnApi.getColumnState()));
+		});
+
+		$scope.gridOptions.api.addEventListener("columnVisible", function() {
+			$scope.gridOptions.api.sizeColumnsToFit();
+			try {
+				colStates = $scope.gridOptions.columnApi.getColumnState();
+				localStorage.setItem(tableName + "_table_columns", JSON.stringify(colStates));
+			} catch (e) {
+				console.error("Failed to store column defs to local storage:", e);
 			}
 		});
 
+		// clicks outside the context menu will hide it
+		$document.bind("click", function(e) {
+			$scope.showMenu = false;
+			e.stopPropagation();
+			$scope.$apply();
+		});
 	});
 
 };
 
-TableDeliveryServicesRequestsController.$inject = ['dsRequests', '$scope', '$state', '$uibModal', '$anchorScroll', '$q', '$location', 'dateUtils', 'locationUtils', 'typeService', 'deliveryServiceService', 'deliveryServiceRequestService', 'messageModel', 'propertiesModel', 'userModel'];
+TableDeliveryServicesRequestsController.$inject = ['tableName', 'dsRequests', '$scope', '$state', '$uibModal', '$anchorScroll', '$q', '$location', '$document', 'dateUtils', 'locationUtils', 'typeService', 'deliveryServiceService', 'deliveryServiceRequestService', 'messageModel', 'propertiesModel', 'userModel'];
 module.exports = TableDeliveryServicesRequestsController;